Fixed fetchAllArticles not fetching the articles correctly
Fixed an issue where fetchAllArticles would always return null because
of an incorrect indexing of the response object in getLanglinks.

Fixed an issue where getLanglinks didn't include the current language.

Fixed an issue where getCurrentUrl would return null when using the
devtools. Now it correctly returns the url of the current window.